Merge eunit and eunit.test tasks
When we use apps_paths, we get the list of apps in an arbitrary order,
which means that we can't reliably compile dependent apps in the correct
order.

This means that -include_lib doesn't work reliably.

So: move the responsibility for compiling stuff in the correct order
elsewhere.

That is:

    MIX_ENV=test mix do compile, eunit

So 'compile' compiles everything in the right order, and 'eunit' _only_
runs the tests.